Add caching options to parallel_region_renewables and add_renewables_clusters functions

+++ powergenome/nrelatb.py
@@ -715,6 +715,8 @@ def parallel_region_renewables(
         region,
         copy.deepcopy(settings),
         cluster_builder,
+        cache_results=True,
+        use_cache=True,
     )
 
     return _df
@@ -801,6 +803,8 @@ def add_renewables_clusters(
     region: str,
     settings: dict,
     cluster_builder: ClusterBuilder = None,
+    cache_results: bool = False,
+    use_cache: bool = False,
 ) -> pd.DataFrame:
     """
     Add renewables clusters
@@ -956,29 +960,34 @@ def add_renewables_clusters(
         cache_folder = Path(
             settings["input_folder"] / "cluster_assignments" / sub_folder
         )
-        add_row_to_csv(
-            cache_folder / "hash_map.csv",
-            headers=[
-                "name",
-                "hash",
-                "metadata_path",
-                "profiles_path",
-                "metadata_sha256",
-                "profiles_sha256",
-            ],
-            new_row=[
-                f"{region}_{technology}_{detail_suffix}_UTC{settings.get('utc_offset', 0)}_file_{data_file_hash}",
-                unique_hash,
-                str(metadata_path),
-                str(profiles_path),
-                metadata_hash,
-                profiles_hash,
-            ],
-        )
+        if cache_results:
+            add_row_to_csv(
+                cache_folder / "hash_map.csv",
+                headers=[
+                    "name",
+                    "hash",
+                    "metadata_path",
+                    "profiles_path",
+                    "metadata_sha256",
+                    "profiles_sha256",
+                ],
+                new_row=[
+                    f"{region}_{technology}_{detail_suffix}_UTC{settings.get('utc_offset', 0)}_file_{data_file_hash}",
+                    unique_hash,
+                    str(metadata_path),
+                    str(profiles_path),
+                    metadata_hash,
+                    profiles_hash,
+                ],
+            )
         cache_cluster_fpath = cache_folder / cache_cluster_fn
         cache_site_assn_fpath = cache_folder / cache_site_assn_fn
         if precluster is False:
-            if cache_cluster_fpath.exists() and cache_site_assn_fpath.exists():
+            if (
+                cache_cluster_fpath.exists()
+                and cache_site_assn_fpath.exists()
+                and use_cache
+            ):
                 clusters = pd.read_parquet(cache_cluster_fpath)
                 data = pd.read_parquet(cache_site_assn_fpath)
             else:
@@ -1015,9 +1024,9 @@ def add_renewables_clusters(
                 )
 
                 cache_folder.mkdir(parents=True, exist_ok=True)
-                if not cache_cluster_fpath.exists():
+                if not cache_cluster_fpath.exists() and cache_results:
                     clusters.to_parquet(cache_cluster_fpath)
-                if not cache_site_assn_fpath.exists():
+                if not cache_site_assn_fpath.exists() and cache_results:
                     cols = ["cpa_id", "cluster"]
                     data[cols].to_parquet(cache_site_assn_fpath)
             if settings.get("extra_outputs"):
@@ -1029,7 +1038,7 @@ def add_renewables_clusters(
                     Path(settings["extra_outputs"]) / fn, index=False
                 )
         else:
-            if cache_cluster_fpath.exists():
+            if cache_cluster_fpath.exists() and use_cache:
                 clusters = pd.read_parquet(cache_cluster_fpath)
                 data = None
             else:
@@ -1047,9 +1056,9 @@ def add_renewables_clusters(
                 clusters["cluster"] = range(1, 1 + len(clusters))
                 data = None
         cache_folder.mkdir(parents=True, exist_ok=True)
-        if not cache_cluster_fpath.exists():
+        if not cache_cluster_fpath.exists() and cache_results:
             clusters.to_parquet(cache_cluster_fpath)
-        if not cache_site_assn_fpath.exists() and data is not None:
+        if not cache_site_assn_fpath.exists() and data is not None and cache_results:
             cols = ["cpa_id", "cluster"]
             data[cols].to_parquet(cache_site_assn_fpath)
         if _scenario.get("min_capacity"):
